+++ lib/kiq.ex
@@ -242,6 +242,46 @@ defmodule Kiq do
 
   [expi]: https://github.com/mperham/sidekiq/wiki/Pro-Expiring-Jobs
 
+  ## Periodic Jobs
+
+  Kiq supports Sidekiq Enterprise's [Periodic Jobs][peri]. This allows jobs to
+  be registered with a schedule and enqueued automatically. Jobs are registered
+  as `{crontab, worker}` or `{crontab, worker, options}` using the `:periodics`
+  attribute:
+
+      use Kiq, periodics: [
+        {"* * * * *", MyApp.MinuteWorker},
+        {"0 * * * *", MyApp.HourlyWorker},
+        {"0 0 * * *", MyApp.DailyWorker, retry: 1},
+      ]
+
+  These jobs would be executed as follows:
+
+    * `MyApp.MinuteWorker` - Executed once every minute
+    * `MyApp.HourlyWorker` - Executed at the first minute of every hour
+    * `MyApp.DailyWorker` - Executed at midnight every day
+
+  The crontab format, as parsed by `Kiq.Parser.Crontab`, respects all [standard
+  rules][cron] and has one minute resolution. That means it isn't possible to
+  enqueue a job ever N seconds.
+
+  #### Notes
+
+    * All schedules are evaluated as UTC, the local timezone is never taken
+      into account.
+    * Periodic jobs registered in Kiq _aren't_ visible in the Loop panel of the
+      Sidekiq Dashboard. This is due to the way loop data is stored by Sidekiq
+      and can't be worked around.
+    * This is an alternative to using using a separate scheduler such as
+      [Quantum][quan]. However, unlike Quantum, Kiq doesn't support node based
+      clustering, instead it uses Redis to coordinate and distrubte work. This
+      means workers can scale horizontally even in a restricted environment like
+      Heroku.
+
+  [peri]: https://github.com/mperham/sidekiq/wiki/Ent-Periodic-Jobs
+  [cron]: https://en.wikipedia.org/wiki/Cron#Overview
+  [quan]: https://github.com/quantum-elixir/quantum-core
+
   ## Instrumentation & Metrics
 
   The instrumentation reporter provides integration with [Telemetry][tele], a